+++ crates/rslint_core/src/groups/style/mod.rs
@@ -1,0 +1,9 @@
+//! Rules which relate to code style and formatting.
+
+use crate::group;
+
+group! {
+    /// Rules which relate to code style and formatting.
+    style,
+    block_spacing::BlockSpacing
+}