+++ crates/rslint_core/src/util/style.rs
@@ -8,19 +8,19 @@ use SyntaxKind::*;
 /// Extensions to nodes and tokens used for stylistic linting which cares about trivia (whitespace)
 pub trait StyleExt {
     fn to_elem(&self) -> SyntaxElement;
-    fn trailing_trivia_has_linebreak(&self) -> bool {
-        self.trailing_whitespace()
+    fn trailing_trivia_has_linebreak(&self, stop_at_comment: bool) -> bool {
+        self.trailing_whitespace(stop_at_comment)
             .into_iter()
             .any(|x| parseutil::contains_js_linebreak(x.text().as_str()))
     }
 
-    fn leading_trivia_has_linebreak(&self) -> bool {
-        self.leading_whitespace()
+    fn leading_trivia_has_linebreak(&self, stop_at_comment: bool) -> bool {
+        self.leading_whitespace(stop_at_comment)
             .into_iter()
             .any(|x| parseutil::contains_js_linebreak(x.text().as_str()))
     }
 
-    fn trailing_trivia(&self) -> Vec<SyntaxToken> {
+    fn trailing_trivia(&self, stop_at_comment: bool) -> Vec<SyntaxToken> {
         let elem = self.to_elem();
         let token = elem
             .clone()
@@ -29,7 +29,7 @@ pub trait StyleExt {
 
         iter::successors(
             token.and_then(|x| {
-                if x.kind().is_trivia() {
+                if x.kind() == WHITESPACE || (x.kind() == COMMENT && !stop_at_comment) {
                     Some(x)
                 } else {
                     x.next_token()
@@ -41,14 +41,14 @@ pub trait StyleExt {
         .collect()
     }
 
-    fn trailing_whitespace(&self) -> Vec<SyntaxToken> {
-        self.trailing_trivia()
+    fn trailing_whitespace(&self, stop_at_comment: bool) -> Vec<SyntaxToken> {
+        self.trailing_trivia(stop_at_comment)
             .into_iter()
-            .filter(|x| x.kind() == WHITESPACE)
+            .take_while(|x| x.kind() == WHITESPACE)
             .collect()
     }
 
-    fn leading_trivia(&self) -> Vec<SyntaxToken> {
+    fn leading_trivia(&self, stop_at_comment: bool) -> Vec<SyntaxToken> {
         let elem = self.to_elem();
         let token = elem
             .clone()
@@ -69,7 +69,7 @@ pub trait StyleExt {
 
         iter::successors(
             token.and_then(|x| {
-                if x.kind().is_trivia() {
+                if x.kind() == WHITESPACE || (x.kind() == COMMENT && !stop_at_comment) {
                     Some(x)
                 } else {
                     x.prev_token()
@@ -81,11 +81,21 @@ pub trait StyleExt {
         .collect()
     }
 
-    fn leading_whitespace(&self) -> Vec<SyntaxToken> {
-        self.leading_trivia()
+    fn leading_whitespace(&self, stop_at_comment: bool) -> Vec<SyntaxToken> {
+        self.leading_trivia(stop_at_comment)
             .into_iter()
-            .filter(|x| x.kind() == WHITESPACE)
+            .take_while(|x| x.kind() == WHITESPACE)
             .collect()
+    }
+
+    fn has_leading_whitespace(&self, include_linebreaks: bool, stop_at_comment: bool) -> bool {
+        let leading = self.leading_whitespace(stop_at_comment);
+        tokens_have_whitespace(&leading, include_linebreaks)
+    }
+
+    fn has_trailing_whitespace(&self, include_linebreaks: bool, stop_at_comment: bool) -> bool {
+        let trailing = self.trailing_whitespace(stop_at_comment);
+        tokens_have_whitespace(&trailing, include_linebreaks)
     }
 }
 
@@ -105,4 +115,15 @@ impl StyleExt for SyntaxElement {
     fn to_elem(&self) -> SyntaxElement {
         self.to_owned()
     }
+}
+
+pub fn tokens_have_whitespace(tokens: &[SyntaxToken], include_linebreaks: bool) -> bool {
+    for ws in tokens {
+        if parseutil::contains_js_whitespace(ws.text().as_str())
+            || (include_linebreaks && parseutil::contains_js_linebreak(ws.text().as_str()))
+        {
+            return true;
+        }
+    }
+    false
 }